Update TextBox.cs accordingly to the code review
The finally with Marshal.FreeBSTR(textPtr) is there because I want to release the pointer (textPtr) no matter what. It's a Win32 pointer and Garbage Collector doesn't release it. It should not be where the object is created because this means I would release it immediately after I create it and before I use it.